Split settings: global app settings / current DAO
Allows the app settings to be loaded without having to wait for the apps
to load. Give to the user the possibility to change the node or reset
the cache if the apps are not loading for some reason.
