+++ src/apps/Settings/Note.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+const Note = styled.p`
+  margin-top: 10px;
+  font-size: 13px;
+  strong {
+    font-weight: 600;
+  }
+`
+
+export default Note
