+++ src/apps/Settings/Settings.js
@@ -1,112 +1,50 @@
 import React from 'react'
 import styled from 'styled-components'
-import {
-  Button,
-  DropDown,
-  Field,
-  TextInput,
-  Text,
-  observe,
-  theme,
-} from '@aragon/ui'
+import { DropDown, Button, Field, TextInput } from '@aragon/ui'
 import AppLayout from '../../components/AppLayout/AppLayout'
-import Option from './components/Option'
-import observeCache from '../../components/HOC/observeCache'
-import EtherscanLink from '../../components/Etherscan/EtherscanLink'
-import provideNetwork from '../../context/provideNetwork'
-import { sanitizeNetworkType } from '../../network-config'
-import { compose } from '../../utils'
-import { getWeb3 } from '../../web3-utils'
-import { web3Providers, network, appIds } from '../../environment'
 import {
   getDefaultEthNode,
   getIpfsGateway,
+  getSelectedCurrency,
   setDefaultEthNode,
   setIpfsGateway,
+  setSelectedCurrency,
 } from '../../local-settings'
-import airdrop, { testTokensEnabled } from '../../testnet/airdrop'
-
-// const AVAILABLE_CURRENCIES = ['USD', 'EUR', 'GBP', 'CAD', 'RMB', 'JPY']
-const AVAILABLE_CURRENCIES = ['USD'] // Only use USD for now
-
-const CACHE_KEY = 'settings'
+import DaoSettings from './DaoSettings'
+import Option from './Option'
+import Note from './Note'
 
 const Content = styled.div`
   max-width: 600px;
   padding: 30px;
 `
 
-const LinkButton = styled(Button.Anchor).attrs({
-  compact: true,
-  mode: 'outline',
-})`
-  background: ${theme.contentBackground};
-`
+// Only USD for now
+const AVAILABLE_CURRENCIES = ['USD']
 
-const AppsList = styled.ul`
-  list-style: none;
-`
-
-const FieldTwoParts = styled.div`
-  display: flex;
-  align-items: center;
-  input {
-    margin-right: 10px;
-    padding-top: 4px;
-    padding-bottom: 4px;
-  }
-`
-
-const Note = styled.p`
-  margin-top: 10px;
-  font-size: 13px;
-  strong {
-    font-weight: 600;
-  }
-`
+// If the currency isn’t available, get the first available instead.
+const filterCurrency = currency => {
+  currency = currency.toUpperCase()
+  return AVAILABLE_CURRENCIES.indexOf(currency) > -1
+    ? currency
+    : AVAILABLE_CURRENCIES[0]
+}
 
 class Settings extends React.Component {
   static defaultProps = {
+    daoAddr: '',
     account: '',
     apps: [],
-    currencies: [],
-    network: '',
   }
   state = {
     defaultEthNode: getDefaultEthNode(),
     ipfsGateway: getIpfsGateway(),
+    currencies: AVAILABLE_CURRENCIES,
+    selectedCurrency: filterCurrency(getSelectedCurrency()),
   }
-  constructor(props) {
-    super(props)
-
-    this.correctCurrencyIfNecessary(props)
-  }
-  componentWillReceiveProps(nextProps) {
-    this.correctCurrencyIfNecessary(nextProps)
-  }
-  correctCurrencyIfNecessary({ cache, currencies, selectedCurrency }) {
-    if (
-      Array.isArray(currencies) &&
-      currencies.indexOf(selectedCurrency) === -1
-    ) {
-      // Oops, somehow the selected currency isn't in the available currencies
-      // Let's reset it to the first available currency
-      this.propagateSelectedCurrency(cache, currencies[0])
-    }
-  }
-  handleCurrencyChange = index => {
-    const { cache, currencies } = this.props
-    this.propagateSelectedCurrency(cache, currencies[index])
-  }
-  propagateSelectedCurrency(cache, selectedCurrency) {
-    cache.update(CACHE_KEY, settings => ({ ...settings, selectedCurrency }))
-  }
-  handleDepositTestTokens = () => {
-    const { account, apps } = this.props
-    const finance = apps.find(app => app.appId === appIds.Finance)
-    if (finance && finance.appId) {
-      airdrop(getWeb3(web3Providers.wallet), finance.appId, account)
-    }
+  handleSelectedCurrencyChange = (index, currencies) => {
+    setSelectedCurrency(currencies[index])
+    this.setState({ selectedCurrency: currencies[index] })
   }
   handleDefaultEthNodeChange = event => {
     this.setState({ defaultEthNode: event.target.value })
@@ -126,154 +64,40 @@ class Settings extends React.Component {
     window.location.reload()
   }
   render() {
+    const { daoAddr, account, apps } = this.props
     const {
+      defaultEthNode,
+      ipfsGateway,
       currencies,
-      daoAddr,
-      account,
-      network: userNetwork,
       selectedCurrency,
-      apps,
-    } = this.props
-    const { defaultEthNode, ipfsGateway } = this.state
-
-    const enableTransactions = !!account && userNetwork === network.type
-    const financeApp = apps.find(({ name }) => name === 'Finance')
+    } = this.state
     return (
       <AppLayout title="Settings">
         <Content>
-          <Option
-            name="Organization address"
-            text={`This organization is deployed on the ${network.name}.`}
-          >
-            <Field label="Address">
-              <FieldTwoParts>
-                <TextInput readOnly wide value={daoAddr} />
-                <EtherscanLink address={daoAddr}>
-                  {url =>
-                    url ? (
-                      <LinkButton href={url} target="_blank">
-                        See on Etherscan
-                      </LinkButton>
-                    ) : null
-                  }
-                </EtherscanLink>
-              </FieldTwoParts>
-              <Note>
-                <strong>Do not send ether or tokens to this address!</strong>{' '}
-                Follow the instructions in the “Send ether to this organization”
-                section below.
-              </Note>
-            </Field>
-          </Option>
-          {testTokensEnabled(network.type) && (
-            <Option
-              name="Request test tokens"
-              text="Deposit some tokens into your organization for testing purposes."
-            >
-              <div>
-                <Button
-                  mode="secondary"
-                  onClick={this.handleDepositTestTokens}
-                  disabled={!enableTransactions}
-                  style={{ opacity: enableTransactions ? 1 : 0.6 }}
-                >
-                  Request test tokens
-                </Button>
-                {!enableTransactions && (
-                  <Text size="small" style={{ marginLeft: '10px' }}>
-                    {(() =>
-                      userNetwork !== network.type
-                        ? `Please select the ${sanitizeNetworkType(
-                            network.type
-                          )} network in MetaMask.`
-                        : `Please unlock your account in MetaMask.`)()}
-                  </Text>
-                )}
-              </div>
-              <Note>
-                Requesting tokens will assign random <strong>TEST</strong>{' '}
-                tokens to your organization. The tokens are named after existing
-                projects, but keep in mind <strong>THEY ARE NOT</strong> the
-                real ones. You can view the received tokens in the Token
-                Balances on the Finance app.
-              </Note>
-            </Option>
-          )}
-          {financeApp && (
-            <Option
-              name="Send ether to this organization"
-              text={
-                <span>
-                  To send <strong>test ether</strong> to this organization, use
-                  the address of the Finance app indicated below.
-                </span>
-              }
-            >
-              <Field label="Funding Address (Finance App)">
-                <FieldTwoParts>
-                  <TextInput readOnly wide value={financeApp.appId} />
-                  <EtherscanLink address={financeApp.appId}>
-                    {url =>
-                      url ? (
-                        <LinkButton href={url} target="_blank">
-                          See on Etherscan
-                        </LinkButton>
-                      ) : null
-                    }
-                  </EtherscanLink>
-                </FieldTwoParts>
-                <Note>
-                  Remember to use the {network.name} network.{' '}
-                  <strong>DO NOT send real ether from the main network!</strong>
-                </Note>
-              </Field>
-            </Option>
-          )}
-          {apps.length > 0 && (
-            <Option
-              name="Aragon apps"
-              text={`This organization has ${apps.length} apps installed.`}
-            >
-              <AppsList>
-                {apps.map(({ name, appId, description }) => (
-                  <li title={description} key={appId}>
-                    <Field label={name}>
-                      <FieldTwoParts>
-                        <TextInput readOnly wide value={appId} />
-                        <EtherscanLink address={appId}>
-                          {url =>
-                            url ? (
-                              <LinkButton href={url} target="_blank">
-                                See on Etherscan
-                              </LinkButton>
-                            ) : null
-                          }
-                        </EtherscanLink>
-                      </FieldTwoParts>
-                    </Field>
-                  </li>
-                ))}
-              </AppsList>
-            </Option>
-          )}
+          <DaoSettings daoAddr={daoAddr} apps={apps} />
           {currencies.length > 1 &&
             selectedCurrency && (
               <Option
                 name="Currency"
-                text="This will be the default currency for displaying purposes. It will be converted to ETH under the hood."
+                text={`
+                  This will be the default currency for displaying purposes.
+                  It will be converted to ETH under the hood.
+                `}
               >
                 <Field label="Select currency">
                   <DropDown
                     active={currencies.indexOf(selectedCurrency)}
                     items={currencies}
-                    onChange={this.handleCurrencyChange}
+                    onChange={this.handleSelectedCurrencyChange}
                   />
                 </Field>
               </Option>
             )}
           <Option
             name="Node settings (advanced)"
-            text="Change which Ethereum and IPFS clients this app is connected to"
+            text={`
+              Change which Ethereum and IPFS clients this app is connected to
+            `}
           >
             <Field label="Ethereum node">
               <TextInput
@@ -295,7 +119,10 @@ class Settings extends React.Component {
           </Option>
           <Option
             name="Troubleshooting"
-            text="Press this button to refresh the cache of the application in your browser."
+            text={`
+              Press this button to refresh the cache of the application in your
+              browser.
+            `}
           >
             <div>
               <Button mode="secondary" onClick={this.handleRefreshCache}>
@@ -313,16 +140,5 @@ class Settings extends React.Component {
     )
   }
 }
-const enhance = compose(
-  observeCache(CACHE_KEY, {
-    defaultValue: {
-      selectedCurrency: AVAILABLE_CURRENCIES[0],
-    },
-    forcedValue: {
-      currencies: AVAILABLE_CURRENCIES,
-    },
-  }),
-  observe(observable => observable.map(settings => ({ ...settings })), {}),
-  provideNetwork
-)
-export default enhance(Settings)
+
+export default Settings
